+++ webofneeds/won-owner-webapp/src/main/webapp/app/service/linkeddata-service-won.js
@@ -17,8 +17,7 @@
 /**
  * Created by fkleedorfer on 05.09.2014.
  */
-import { is, clone, getIn, get, getInFromJsonLd } from "../utils.js";
-import * as wonUtils from "../won-utils.js";
+import { is, clone, getIn, get } from "../utils.js";
 
 import { ownerBaseUrl } from "~/config/default.js";
 import urljoin from "url-join";
@@ -1808,41 +1807,3 @@ function rethrow(e, prependedMsg = "") {
     throw new Error(prependedMsg + JSON.stringify(e));
   }
 }
-
-/**
- * Traverses the `path` into the json-ld object and then tries to
- * parse that node as RDF literal object (see `wonUtils.parseJsonldLeafsImm` for
- * details on that).
- *
- * @param {*} jsonld
- * @param {*} path
- * @param {*} type optional (see getFromJsonLd for details)
- * @param {*} context defaults to `won.defaultContext`
- * @return the list at the path
- */
-won.parseListFrom = (jsonld, path, type, context = won.defaultContext) => {
-  try {
-    return wonUtils.parseJsonldLeafsImm(
-      getInFromJsonLd(jsonld, path, context),
-      type
-    );
-  } catch (err) {
-    console.error("Could not parse From list: ", err);
-    return undefined;
-  }
-};
-
-/**
- * Traverses the `path` into the json-ld object and then tries to
- * parse that node as RDF literal object (see `wonUtils.parseJsonldLeaf` for
- * details on that).
- *
- * @param {*} jsonld
- * @param {*} path
- * @param {*} type optional (see getFromJsonLd for details)
- * @param {*} context defaults to `won.defaultContext`
- * @return the value at the path
- */
-won.parseFrom = (jsonld, path, type, context = won.defaultContext) => {
-  return wonUtils.parseJsonldLeaf(getInFromJsonLd(jsonld, path, context), type);
-};
