+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-complain.js
@@ -2,14 +2,14 @@
  * Created by fsuda on 18.09.2018.
  */
 import { details, mergeInEmptyDraft } from "../detail-definitions.js";
-import * as wonUtils from "../../app/won-utils.js";
+import * as jsonLdUtils from "../../app/service/jsonld-utils.js";
 
 export const complain = {
   identifier: "complain",
   label: "Complain about something",
   icon: "#ico36_uc_wtf",
   timeToLiveMillisDefault: 1000 * 60 * 60 * 24 * 30,
-  doNotMatchAfter: wonUtils.findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+  doNotMatchAfter: jsonLdUtils.findLatestIntervallEndInJsonLdOrNowAndAddMillis,
   draft: {
     ...mergeInEmptyDraft({
       content: {
