+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-custom.js
@@ -1,11 +1,11 @@
 import { details, mergeInEmptyDraft } from "../detail-definitions.js";
-import * as wonUtils from "../../app/won-utils.js";
+import * as jsonLdUtils from "../../app/service/jsonld-utils.js";
 
 export const customUseCase = {
   identifier: "customUseCase",
   label: "New custom post",
   icon: "#ico36_uc_custom",
-  doNotMatchAfter: wonUtils.findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+  doNotMatchAfter: jsonLdUtils.findLatestIntervallEndInJsonLdOrNowAndAddMillis,
   draft: { ...mergeInEmptyDraft() },
   details: details,
   seeksDetails: details,
