+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-job-search.js
@@ -1,6 +1,6 @@
 import { details, mergeInEmptyDraft } from "../detail-definitions.js";
 import { interestsDetail, skillsDetail } from "../details/person.js";
-import * as wonUtils from "../../app/won-utils.js";
+import * as jsonLdUtils from "../../app/service/jsonld-utils.js";
 import {
   industryDetail,
   employmentTypesDetail,
@@ -24,7 +24,7 @@ export const jobSearch = {
   identifier: "jobSearch",
   label: "Search a Job",
   icon: "#ico36_uc_consortium-search", //TODO proper icon
-  doNotMatchAfter: wonUtils.findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+  doNotMatchAfter: jsonLdUtils.findLatestIntervallEndInJsonLdOrNowAndAddMillis,
   draft: {
     ...mergeInEmptyDraft({
       content: {
