+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-rehearsal-room-offer.js
@@ -9,7 +9,7 @@ import {
 } from "../details/real-estate.js";
 import won from "../../app/won-es6.js";
 import { perHourRentDetail } from "../details/musician.js";
-import * as wonUtils from "../../app/won-utils.js";
+import * as jsonLdUtils from "../../app/service/jsonld-utils.js";
 import {
   filterNumericProperty,
   filterPrice,
@@ -22,7 +22,7 @@ export const rehearsalRoomOffer = {
   label: "Offer Rehearsal Room",
   icon: "#ico36_uc_realestate",
   timeToLiveMillisDefault: 1000 * 60 * 60 * 24 * 30 * 3,
-  doNotMatchAfter: wonUtils.findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+  doNotMatchAfter: jsonLdUtils.findLatestIntervallEndInJsonLdOrNowAndAddMillis,
   draft: {
     ...mergeInEmptyDraft({
       content: {
